Move iterators into separate file
